Tests for MateController

+++ server/src/main/java/server/api/MateController.java
@@ -1,10 +1,14 @@
 package server.api;
 
+import commons.House;
 import commons.Mate;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.web.bind.annotation.*;
+import server.database.HouseRepository;
 import server.database.MateRepository;
 
 import java.util.List;
@@ -13,9 +17,12 @@ import java.util.List;
 @RequestMapping("/api/mates")
 public class MateController {
     private final MateRepository repo;
+    @Autowired
+    HouseRepository houseRepo;
 
-    public MateController(MateRepository repo) {
+    public MateController(MateRepository repo, HouseRepository houseRepo) {
         this.repo = repo;
+        this.houseRepo = houseRepo;
     }
 
     @GetMapping(path = {"", "/"})
@@ -31,15 +38,24 @@ public class MateController {
         return ResponseEntity.ok(repo.findById(id).get());
     }
 
-    @MessageMapping({"/mates/add"})
+    @MessageMapping({"/mates/add/{houseId}"})
     @SendTo("/topic/mates/update")
-    public Mate addMessage(Mate mate) {
-        return add(mate).getBody();
+    public Mate addMessage(Mate mate, @DestinationVariable long houseId) {
+        return add(mate, houseId).getBody();
     }
 
-    @PostMapping(path = {"", "/"})
-    public ResponseEntity<Mate> add(@RequestBody Mate mate) {
+    @PostMapping({"add/{houseId}"})
+    public ResponseEntity<Mate> add(
+            @RequestBody Mate mate,
+            @PathVariable long houseId
+    ) {
+        if (mate == null)
+            return ResponseEntity.badRequest().build();
+
+        House assoc = houseRepo.getById(houseId);
         Mate saved = repo.save(mate);
+        saved.setHouse(assoc);
+        saved = repo.save(saved);
         return ResponseEntity.ok(saved);
     }
 
@@ -52,7 +68,38 @@ public class MateController {
 
     @DeleteMapping(path = {"", "/"})
     public ResponseEntity<Mate> delete(@RequestBody Mate mate) {
+        if (mate == null
+                || !repo.existsById(mate.getId()))
+            return ResponseEntity.badRequest().build();
+
         repo.delete(mate);
         return ResponseEntity.ok(mate);
     }
+
+    /*
+    @MessageMapping("/mate/replace")
+    @SendTo("/topic/list/replace")
+    public Mate replaceMessage(Mate mate) {
+        return replace(mate, mate.getId()).getBody();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Mate> replace(
+            @RequestBody Mate mate,
+            @PathVariable("id") long id
+    ) {
+        if (mate == null
+                || !repo.existsById(id))
+            return ResponseEntity.badRequest().build();
+
+        Mate mateToChange = repo.findById(id).isPresent() ? repo.findById(id).get() : null;
+
+        if (mateToChange != null) {
+            delete(mate);
+            add(mate, mateToChange.getHouse().getId());
+            mateToChange.setHouse(null);
+        }
+
+        return ResponseEntity.ok(mateToChange);
+    }*/
 }
